requirements.txt
Flask==3.0.0
python-pptx==0.6.23